+++ R/pkgJSON.R
@@ -47,10 +47,9 @@ createJSON <- function(repo, pkg_name, descr_df, scm_df, docdir,
 
   # Convert to JSON
   desc_json <- toJSON(descr_df, pretty = TRUE)
-  # Make sure JSON is not in the form of a list,
-  # Trim leading and trailing list markers i.e. '[' and ']'
+  # Remove leading and lagging JSON list markers i.e. '[' and ']'
   desc_json <- substring(desc_json, 2)
-  desc_json <- substr(desc_json, 1, nchar(desc_json) - 1)
+  desc_json <- substr(desc_json, 1, nchar(desc_json)-1)
 
   # Write JSON
   json_outfile <- file.path(docdir, paste0(pkg_name, suffix))