+++ R/saveRepo.R
@@ -20,7 +20,7 @@ saveRepoFiles <- function(repo) {
 }
 
 finalizeRepo <- function(repo) {
-  manifestHTML(repo)
+  buildReport(repo)
   saveRepoFiles(repo)
   if (!file.exists(file.path(repobase(repo), "getGRAN.R"))) {
     file.copy(file.path(repobase(repo), paste0("GRAN", repo_name(repo)), "inst",